Do not assume all pair_coeff come with a path
At least for MTP potentials the format for the lammps potential is like

pair_style mlip mlip.ini
pair_coeff * *

where mlip.ini is the potential specific file.  That means when
fix_paths tries to look a filepath to adjust in pair_coeff it won't find
anything and crash.  This also means fix_paths might have to be applied
to pair_style also, but I haven't done that here. :/